+++ src/tokenizer/number_as_string.rs
@@ -5,25 +5,25 @@ use std::ops::Range;
 #[derive(Debug, Clone)]
 pub struct Number {
     /// The byte-range of the original substring in the input.
-    str_range: Range<u32>,
+    str_range: Range<usize>,
 }
 
 impl Number {
     /// Constructs a `Number` by capturing the range of characters to treat as a string.
     #[inline]
-    pub fn new(_input: &str, range: Range<u32>) -> Self {
+    pub fn new(_input: &str, range: Range<usize>) -> Self {
         Number { str_range: range }
     }
 
     /// Returns the original substring slice from the input.
     #[inline]
     pub fn as_bytes<'a>(&self, input: &'a str) -> &'a [u8] {
-        &input.as_bytes()[self.str_range.start as usize..self.str_range.end as usize]
+        &input.as_bytes()[self.str_range.start..self.str_range.end]
     }
 
     /// Converts the stored substring into a standalone `String`.
     #[inline]
     pub fn to_string(&self, input: &str) -> String {
-        input[self.str_range.start as usize..self.str_range.end as usize].to_string()
+        input[self.str_range.start..self.str_range.end].to_string()
     }
 }
